Constrained parsing appears to work properly, although it often fails on non-canonical text, so we probably need a recovery mode there too. That will require a more sophisticated recovery than currently permitted by -recovery.

+++ java/src/edu/ohsu/cslu/parser/ml/CartesianProductHashSpmlParser.java
@@ -91,7 +91,7 @@ public class CartesianProductHashSpmlParser extends
                     for (int k = binaryColumnOffsets[column]; k < binaryColumnOffsets[column + 1]; k++) {
 
                         final float jointProbability = binaryProbabilities[k] + childProbability;
-                        final int parent = binaryRowIndices[k];
+                        final short parent = binaryRowIndices[k];
 
                         if (jointProbability > tmpCell.insideProbabilities[parent]) {
                             tmpCell.packedChildren[parent] = column;
